review feedback: move all stat-related fns to root
